новые функции редактирования и удаления

+++ webapp/views.py
@@ -1,12 +1,10 @@
-from django.shortcuts import render, redirect
+from django.shortcuts import render, redirect, get_object_or_404
 from .models import Task
 from .forms import TaskForm
-from django.shortcuts import get_object_or_404, render
 
 def task_list(request):
     tasks = Task.objects.all()
     return render(request, 'task_list.html', {'tasks': tasks})
-
 
 def add_task(request):
     if request.method == 'POST':
@@ -18,13 +16,26 @@ def add_task(request):
         form = TaskForm()
     return render(request, 'add_task.html', {'form': form})
 
-
-def delete_task(request, task_id):
-    task = get_object_or_404(Task, id=task_id)
-    task.delete()
-    return redirect('task_list')
-
-def task_detail(request, task_id):
-    task = get_object_or_404(Task, id=task_id)
+def task_detail(request, pk):
+    task = get_object_or_404(Task, pk=pk)
     return render(request, 'task_detail.html', {'task': task})
+
+def edit_task(request, pk):
+    task = get_object_or_404(Task, pk=pk)
+    if request.method == 'POST':
+        form = TaskForm(request.POST, instance=task)
+        if form.is_valid():
+            form.save()
+            return redirect('task_list')
+    else:
+        form = TaskForm(instance=task)
+    return render(request, 'edit_task.html', {'form': form})
+
+def delete_task(request, pk):
+    task = get_object_or_404(Task, pk=pk)
+    if request.method == 'POST':
+        task.delete()
+        return redirect('task_list')
+    return render(request, 'delete_task.html', {'task': task})
+
